🚀 feat (day 02): Modified into solution for part 2.
Second part of this task changed requirements for validation of passwords.

"Each policy actually describes two positions in the password, where 1 means the first character, 2 means the second character, and so on. ... Exactly one of these positions must contain the given letter. Other occurrences of the letter are irrelevant for the purposes of policy enforcement."